+++ mock/cpp_mock_scenarios/src/lane_change/lanechange_linear_lateral_velocity.cpp
@@ -38,15 +38,19 @@ public:
   }
 
 private:
-  bool requested = false;
-  bool lanechange_finished = false;
+  int lanechange_frames = 0;
   void onUpdate() override
   {
-    if (api_.isInLanelet("ego", 34513, 0.1)) {
-      lanechange_finished = true;
+    if (api_.getCurrentAction("ego") == "lane_change") {
+      lanechange_frames++;
     }
-    if (lanechange_finished && api_.isInLanelet("ego", 34510, 0.1)) {
-      stop(cpp_mock_scenarios::Result::SUCCESS);
+    if (api_.getCurrentAction("ego") != "lane_change" && api_.getCurrentTime() >= 2.0) {
+      double duration = static_cast<double>(lanechange_frames) * 0.05;
+      if (duration >= 3.05 && 3.1 >= duration) {
+        stop(cpp_mock_scenarios::Result::SUCCESS);
+      } else {
+        stop(cpp_mock_scenarios::Result::FAILURE);
+      }
     }
     // LCOV_EXCL_START
     if (api_.getCurrentTime() >= 10.0) {
@@ -67,7 +71,7 @@ private:
         "ego", traffic_simulator::lane_change::Direction::LEFT, 1, 0),
       traffic_simulator::lane_change::TrajectoryShape::LINEAR,
       traffic_simulator::lane_change::Constraint(
-        traffic_simulator::lane_change::Constraint::Type::LONGITUDINAL_DISTANCE, 30.0));
+        traffic_simulator::lane_change::Constraint::Type::LATERAL_VELOCITY, 1.0));
   }
 };