+++ simulation/traffic_simulator/include/traffic_simulator/math/bounding_box.hpp
@@ -23,7 +23,7 @@
 #include <boost/geometry/geometries/point_xy.hpp>
 #include <boost/optional.hpp>
 #include <geometry_msgs/msg/pose.hpp>
-#include <traffic_simulator/math/transfrom.hpp>
+#include <traffic_simulator/math/transform.hpp>
 #include <traffic_simulator_msgs/msg/bounding_box.hpp>
 #include <vector>