+++ simulation/traffic_simulator/include/traffic_simulator/traffic_lights/traffic_light.hpp
@@ -37,9 +37,10 @@ class TrafficLight
 
 public:
   const std::int64_t id;
+  const std::int64_t relation_id;
 
   explicit TrafficLight(
-    std::int64_t id,
+    std::int64_t id, std::int64_t relation_id,
     const std::unordered_map<TrafficLightColor, geometry_msgs::msg::Point> & color_positions = {},
     const std::unordered_map<TrafficLightArrow, geometry_msgs::msg::Point> & arrow_positions = {});