+++ simulation/traffic_simulator/src/entity/entity_base.cpp
@@ -85,7 +85,7 @@ void EntityBase::requestSpeedChange(
 void EntityBase::requestLaneChange(
   const traffic_simulator::lane_change::AbsoluteTarget & target,
   const traffic_simulator::lane_change::TrajectoryShape trajectory_shape,
-  const lane_change::Constraint & constraint)
+  const traffic_simulator::lane_change::Constraint & constraint)
 {
   auto param = traffic_simulator::lane_change::Parameter(target, trajectory_shape, constraint);
   requestLaneChange(param);
@@ -94,7 +94,7 @@ void EntityBase::requestLaneChange(
 void EntityBase::requestLaneChange(
   const traffic_simulator::lane_change::RelativeTarget & target,
   const traffic_simulator::lane_change::TrajectoryShape trajectory_shape,
-  const lane_change::Constraint & constraint)
+  const traffic_simulator::lane_change::Constraint & constraint)
 {
   std::int64_t reference_lanelet_id = 0;
   if (target.entity_name == name) {