+++ docs/ReleaseNotes.md
@@ -4,6 +4,9 @@
 - Start supporting linear trajectory shape while changing lane. ([pull request](https://github.com/tier4/scenario_simulator_v2/pull/661))
 - Fix context panel to display simulation time. (Contribution by [Utaro-M](https://github.com/Utaro-M)).
 - Support new vehicle_model_type `DELAY_STEER_VEL` ([pull request](https://github.com/tier4/scenario_simulator_v2/pull/660)).
+- Add lateral velocity and time constraint. ([pull request](https://github.com/tier4/scenario_simulator_v2/pull/664)).
+- Update syntax `SpeedAction` to use `API::requestSpeedChange` ([pull request](https://github.com/tier4/scenario_simulator_v2/pull/665)).
+- Enable control multiple traffic lights by relation ID from lanelet2 map. ([pull request](https://github.com/tier4/scenario_simulator_v2/pull/667)).
 
 ## Version 0.6.1
 - Add API::requestSpeedChange function. ([link](https://github.com/tier4/scenario_simulator_v2/pull/618))